src/requirements.txt
aniso8601==9.0.1
astroid==2.12.9
attrs==22.1.0
click==8.1.3
dill==0.3.5.1
dnspython==2.2.1
Flask==2.1.2
flask-restx==0.5.1
gunicorn==20.1.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
PyJWT==2.5.0
pylint==2.15.2
pymodm==0.4.3
pymongo==3.12.3
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.4
typing-extensions==4.3.0
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.1
